embedding_service/requirements.txt
fastapi
uvicorn
sentence-transformers
spacy
pymupdf
python-docx
supabase
python-dotenv
apscheduler
python-multipart